Fixed minor warnings and a problem with copy constructor in GraspTarget.
git-svn-id: https://svnsrv.sdu.dk/svn/RobWork/trunk@5135 494c09c7-a67d-49d4-9938-151fb1965be8

+++ RobWorkSim/src/rwsim/rwphysics/ConstraintSolver.cpp
@@ -7,7 +7,7 @@ bool ConstraintSolver::solve( std::vector<CEdgeGroup>& groups,
             SolverInfo& info,
             rw::kinematics::State& state) {
 
-    for(int i=0;i<groups.size();i++){
+    for(int i=0;i<(int)groups.size();i++){
         if( !solveGroup(groups[i], info, state) )
             return false;
     }
